[iOS] Dismiss manage accounts after primary account removal
The view gets dismissed by its parent when viewed from settings or from
the account menu. But not when viewed from web, this CL handles this
case to also dismiss it when viewd from web and the primary identity is
removed.

Bug: 366265284
Change-Id: I4032598a51bfc30c24eec9eb7a7ef0f420cd3e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5865470
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1355856}