[enrollment] Remove fallback to attestation mode in EnrollmentScreen
It used to be possible to get EnrollmentConfig with MODE_NONE and
AUTH_MECHANISM_[ATTESTATION, BEST_AVAILABLE] via a switch.
This configuration was treated as a prescribed enrollment, despite
MODE_NONE.

EnrollmentScreen::SetConfig() handled this case by updating the mode to
MODE_ATTESTATION[_LOCAL_FORCED] to propagate a correct config further.

The switch was removed in crrev.com/c/5453851, and auth mechanism only
means preferred mechanism, and not an enrollment mode.

Remove the check in EnrollmentScreen as it is not triggered anymore and
EnrollmentConfig always has a mode != MODE_NONE at this point.

Bug: b:332529631
Test: CQ
Change-Id: I891ad10502f340b25250428a9803480e0a23d1cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5518350
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Roland Bock <rbock@google.com>
Cr-Commit-Position: refs/heads/main@{#1355860}