[ios] Fix "managed by" with large text sizes
The "managed by" text behaves weird when it overflows the screen. It
stays centered, and expands beyond the screen on either side. Only the
"middle" part of the text is visible, making it unreadable.

Bound the text box to the screen's width, so it uses ellipses instead.

Screenshot:
https://drive.google.com/file/d/13FXrhs0t_VmqTB5w2NUueE6FUINMq9EG/view

Bug: 365172149
Change-Id: Ie551121b0068d97cb80c56864ce6f5ba10332ad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5866310
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355869}