[iOS] Improve settings child coordinator checks
As a follow up to bug 40067451, where it was determined that it
is possible to get into the situation where a child coordinator
already exists when attempting to create a new one, this CL
changes the check strategy for these cases to instead verify that
the top view controller on the navigation stack is the current
view controller (so, SettingsTableViewController for the root
settings page, and the password manager / issue view controller for
the password settings.

Because this situation is still not well understood, it felt more
prudent to use DumpWithoutCrashing for now, at least until the
impact of this change is assessed.

Low-Coverage-Reason: TRIVIAL_CHANGE This CL only adds checks
Bug: 40067451, 41484197
Change-Id: Idb1f687fbe05eaefc2ffe8af1a5a841793a177a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782388
Reviewed-by: Noémie St-Onge <noemies@google.com>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355915}