[iOS] Remove CHECK from feed metrics recorder when changing segment
This CL removes a CHECK statement on NTP feed header so that it would
not trigger a crash.

Ideally, `onSegmentSelected` should never have been triggered when the
NTP coordinator, which acts as the delegate for `followDelegate` for the
FeedMetricsRecorder, is `nil`; nevertheless, it tricky to verify what
actually happens since the superclass of `feedHeaderView` is a google3
component, and is actually a UICollectionView.

Bug: 348353795
Change-Id: I3f6a6cd085c4e552fe21a608742b986fd75cf76d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5865252
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1355932}