Handle blob URL deregister on same-site RFH swap in origin comparison code
Now that blob URL calculations can fully happen on the browser side, we
can compare browser vs renderer origin in all blob cases. One exception
is when we do a same-site RFH swap from the document that registers the
blob URL we're navigating to. In this case the blob URL's origin might
get deleted from the map before the commit finishes, causing the
renderer to think that the origin is opaque. In this case the browser
side has the correct origin, which is used as the precursor for the
renderer's opaque origin. This CL makes us compare the precursor for
this case.

Bug: 888079
Change-Id: I52da40a27457344a588e06305115b57e1597be07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860177
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356018}