[codeql] Add a configuration file for setting which CodeQL queries get run.
In the future there will be only one configuration file for both "build this database" and "run this query", but we need to check in this new configuration file first before we can deprecate the old one.


[codeql] Add a config file to determine what CodeQL queries to run in automation.

This will be utilized by the chrome_codeql_query_runner.py recipe.

Change-Id: I2e391f1c9ec5fa434ce7fea73d065ef9dae7bd9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5862367
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1356023}