privacy-hub: Hide geolocation dialog when pref is managed (part 3)
Users have ability to enable geolocation inline from system settings
that use location. When the geolocation pref is managed, users should
not see this dialog, as well as the text that triggers this dialog.

Handle the Theme section of the Personalization app. This is the final
patch in this series.

screenshots:
- https://screenshot.googleplex.com/36kAfiM3e3NJTFJ
- https://screenshot.googleplex.com/5Ce8cCjnVkzKXkB

Bug: b:360315009
Change-Id: I5886f3e21321dcd28c15451eb409e5ba0d7f877f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5853861
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Jerry Liu <pzliu@google.com>
Commit-Queue: Jeffrey Young <cowmoo@google.com>
Reviewed-by: Jeffrey Young <cowmoo@google.com>
Auto-Submit: Zauri Meshveliani <zauri@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356066}