Allow service to update host_owner email
In order to support email address changes for the user's account,
we need to ensure the host allows connections from the user's
primary email address which is provided by the backend in the
initial heartbeat. Since the original email address is still valid,
and the user can still sign in using it, we will continue to allow
connections to the machine from it as well.

Keeping a list of valid email addresses is straight-forward however
one problem is how to apply Enterprise policies to this list. The
plan is to filter out any email addresses which are denied due to
username or domain policy requirements.  If there are no valid
email addresses after the filtering then the host will shutdown
with an appropriate error message.

Bug: 346825839
Change-Id: I07d0bcc01b0e2905c7efa8e93b89808d428e184f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5867014
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Commit-Queue: Gary Kacmarcik <garykac@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356091}