[ASH] Show IPH only when ASH is enabled.
App-specific history IPH is shown even when the feature is disabled.
The condition for IPH visibility was not the same as that of the
feature itself. This CL fixes it to let them both use the same
condition.

Bug: 365787324
Change-Id: I255397ddb1e71482ac3f4812fdcabd74b5df9b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5849574
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356102}