Speculative fix for dangling ptr crashes on DownloadFileImpl::SourceStream
This is likely to be a false positive because code like this is "safe"
and doesn't crash:
  auto x = std::make_unique<MyClass>();
  auto* p = new int(444);
  auto cb = base::BindOnce(&MyClass::Method, x->GetWeakPtr(), p);
  delete p;
  x.reset();
  // `p` is dangling but this is "okay" since the callback is cancelled,
  // since the `WeakPtr` is a receiver and is invalidated.
  std::move(cb).Run();

However, breaking this up into a nested callback like this:
  auto x = std::make_unique<MyClass>();
  auto* p = new int(444);
  auto inner = base::BindOnce(&MyClass::Method, x->GetWeakPtr());
  auto outer = base::BindOnce(std::move(inner), p);
  delete p;
  x.reset();
  // Crashes in the dangling pointer check.
  std::move(cb).Run();

The special cancellation logic doesn't take effect here, so the callback
implementation must assume that the pointer might be dereferenced and
check if it is dangling.

It is possible that callbacks should check for chained cancellation
here and allow "safe" dangling pointers here as well. However, nesting
callbacks like this is inherently inefficient, so for now, just avoid
nesting callbacks as a fix / workaround.

Bug: 365775467
Change-Id: Ie644ad11049e00e723853c610fd81393af2ecd74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5858326
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356118}