[Payments] Measure usage of a deprecated payment method name
To deprecate the special casing[1] of https://android.com/pay payment
method identifier in PaymentRequest API on Android, add a use counter
for this payment method identifier.

[1]PaymentRequest does not make network requests to
https://android.com/pay on Android, but instead invokes the Google Pay
API directly. However, this payment method identifier has been absent
from documentation for a number of years, as it was replaced with
https://google.com/pay.

This counter aims to confirm that (almost?) no websites are using the
deprecated identifier today and the special casing can be removed.

Bug: 363516222
Change-Id: I31d8571973c04f62422b81c0476badb2c41a74d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5861089
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Darwin Yang <darwinyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356219}