[Settings] Convert back SelectLanguageFragment to a standalone page
LanguageSettings launches SelectLanguageFragment with
startActivityWithResult to receive the user choice and update settings.
Since returning results is possible only by activities, not by
fragments, it is not straightforward to share the same instance of
SettingsActivity between LanguageSettings and SelectLanguageFragment.

Therefore this patch simply converts back SelectLanguageFragment to a
standalone page by unimplementing SettingsPage.

This means that the language list would be shown in a full window even
after we enable the two-column UI, which we might think as suboptimal.
We might want to revisit this after fully enabling the single-activity
mode, with which it becomes easier to pass non-serializable objects,
such as delegates, between fragments because they are guaranteed to
share the same activity instance.

This patch fixes the language settings page in the single-activity mode.

Bug: b/356743945
Change-Id: I9bb53e92f83510ba603869879d4baab8adb091bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5856843
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356248}