Add plumbing for VEAs to create a CommandBufferHelper
In order for VEAs to be able to access shared images, they need to get
a SharedImageManager and (at least on Windows) the same device on which
those textures were created.  To do this, plumb through the GpuChannel's
channel token from where the encoder is created in the renderer, through
mojo, to the factory that creates the VEA implementation.  The VEA can
then use the channel token to create a CommandBufferHelper, giving it
access to SharedImageManager and a shared graphics device.

See http://crrev.com/c/5758929 for how this change fits in to allow
MFVEA to use shared images for encode.

Bug: 40162806
Change-Id: I6e1ea94e23b061004c79dd59b8b8937f07130eed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5800417
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Matthew Andrews <mattandr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1356260}