[text-wrap] Fix a `DCHECK` failure in `LineBreaker::HandleOverflow`
While handling overflow for a hyphenated line. Because
`position_` is changed by a call to `RemoveHyphen`,
`width_to_rewind`, which is dependent on `position_`, should
be computed after the change.

The tests for crbug.com/40943808 hits this `DCHECK`. Tests
will be in the CL for the issue.

Fixed: 366640434
Change-Id: Iaec8602eb6180598d0f4cd82fa78a833a05e60f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5863589
Commit-Queue: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356263}