Remove ScreenAI component Install/Uninstall success metrics.
Component install metric mostly monitors component installer's internal
behavior for non-ChromeOS and is not a representative of the specific
component's availability. "Accessibility.ScreenAI.Component.Available2"
is now used as its replacement to monitor whether the component was
really available when needed or not, and on ChromeOS
"Accessibility.ScreenAI.DlcInstallResult" records the error code when
installation fails.
Uninistallation success is not used for any decision making and hence
collecting it is useless.

OBSOLETE_HISTOGRAMS=Accessibility.ScreenAI.Component.{Action} are not needed anymore.

AX-Relnotes: n/a
Change-Id: Ie9456711ba948339b4bfe05a1bfada894738521c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857169
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356268}