Detangle PickerSuggestionsController lifetime from PickerClient
Right now, PickerSuggestionsController takes a PickerClient in the
ctor, which means when the PickerClient changes or is destroyed,
we have to recreate the PickerSuggestionsController.

This is unnecessary, because the client is only used to delegate
some suggestions, it is not needed as soon as the suggestions
are requested.

Bug: b:360229206
Change-Id: I76370b5efe6d6640c4f40c4b7a78db2d55973487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5869709
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1356297}