[Blink] Make CanvasHibernationHandler test more self-contained
This CL changes the `CanvasWriteWhileCompressing` test to be a
self-contained unittest on `CanvasHibernationHandler` that tests that
invoking `Clear()` while compression is in progress ends hibernation,
and that the subsequent callback that compression is complete is then a
no-op.

Adding an equivalent test on `CanvasRenderingContext2D` that operates at
the level of `putImageData()` rather than
`CanvasHibernationHandler::Clear()` would be a bit fiddly, as we'd have
to add the ability to configure `CanvasHibernationHandler` to fire the
compression task without firing its completion callback. I'm going to
investigate how feasible it is to do that in a way that's not invasive
to the production code. In the meantime, we already have added a test
that invoking `putImageData()` causes hibernation to be ended here [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5860517

Bug: 40280152
Change-Id: I927148b923d7b6ccc755548876acda3431d3ad21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5864311
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356299}