Reduce fragility in SidePanelCoordinator (part 7 / N)
This CL is a refactor with no intended user-visible impact.

This CL deletes the view property SidePanelContentState. This is only
read by SidePanel and only set by SidePanelCoordinator. By removing the
view property we remove an unnecessary layer of indirection. This is
replaced by Open and Close methods on SidePanel which
SidePanelCoordinator calls.

This CL adds a state member to SidePanel to track and expose the
internal state machine. State is one of: kClosed, kOpening, kOpen,
kClosing.

SidePanelCoordinator is the only entity that should be controlling
visibility of SidePanel. All other calls to UpdateVisibility() were
removed. These calls appear to be side effects of relying of
view-properties to propagate state changes.

Bug: 363743081
Change-Id: I7841a40648ba84969eccc9a2448e1283e8caa1d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5858429
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1356309}