Add remaining non-WebContents based GuestView lookups
The last types we need to provide a GuestView lookup method for are
NavigationHandles and FrameTreeNodeIds. We introduce methods for these
and convert a number of call sites off of the WebContents based lookup.

The new methods are still using WebContents internally, but that will
change once the new guest view architecture is introduced.

Bug: 40238394, 40202416
Change-Id: I94adcb04728cdd4b1a7a90513aa694220321d6de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5862048
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Auto-Submit: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356312}