[iOS] Fix password notification crash due to invalid state
This CL addresses a crash in NotificationForPasswordCheckState() caused
by an invalid InsecurePasswordCounts being passed to the function. The
CHECK() that was causing the crash has been removed. Instead, the
function now conditionally returns notification content only if both
inputs are valid.

It's possible that the InsecurePasswordCounts are empty if the function
is called early in the application lifecycle, before the insecure
password counts are refreshed from cache. Instead of CHECK()ing in this
case, the function now gracefully returns nil.

This is acceptable because Safety Check Notifications are potentially
scheduled every time the app is foregrounded, or a new tab page is
opened. It's OK if this helper function is called and infrequently
returns nil because on subsequent foregrounding or new tab page opening,
the correct state will be available to generate appropriate notification
content.

Fixed: 365530176
Change-Id: Ic5937cd9384a7546df089e672f100787e328e0cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5866974
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Louis Romero <lpromero@google.com>
Auto-Submit: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1356320}